requirements.txt
blinker==1.9.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
folium==0.19.5
haversine==2.9.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.4
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2025.1.0
gunicorn